% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{enterOutputPath}
\alias{enterOutputPath}
\title{enterOutputPath}
\usage{
enterOutputPath(path)
}
\arguments{
\item{path}{Enter path where you want to store output}
}
\value{
path for output as a named object called outputDir
}
\description{
enterOutputPath
}
\examples{
enterOutputPath(getwd())
}
