% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults()
}
\value{
A data frame with all ntlx results located in the directory specified
in the enterOutputPath call
}
\description{
getResults
}
\examples{
getResults()
}
