% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processResults.R
\name{processResults}
\alias{processResults}
\title{processResults}
\usage{
processResults()
}
\value{
data frame with raw scores, weights, adjusted scores, and overall weighted workload score for each subject x task combination
}
\description{
processResults
}
\examples{
processResults()
}
